+++ comprehensive_demo_project/src/comprehensive_demo/__init__.py
@@ -11,3 +11,4 @@ __all__ = ["create_recommendation_task"]
 
 
 
+