+++ comprehensive_demo_project/src/comprehensive_demo/task_registry.py
@@ -15,3 +15,4 @@ __all__ = ["recommendation_task"]
 
 
 
+